+++ hnote/actions/types.ts
@@ -1,5 +1,7 @@
 export const ADD_LIST = 'ADD_LIST';
 
+export const DELETE_LIST = 'DELETE_LIST';
+
 export const ADD_TASK_TO_LIST = 'ADD_TASK_TO_LIST';
 
 export const REMOVE_TASK_FROM_LIST = 'REMOVE_TASK_FROM_LIST';